refactor tool_call to function_call; vllm health endpoint
