+++ functionary/sglang_inference.py
@@ -41,6 +41,7 @@ from functionary.inference_stream import generate_openai_format_from_stream_asyn
 from functionary.inference_utils import (
     analyze_tools_and_tool_choice,
     check_all_errors,
+    convert_tool_calls_to_function_call,
     create_error_response,
 )
 from functionary.openai_types import (
@@ -81,25 +82,6 @@ class ChatCompletionParams:
     tool_func_choice: Optional[Union[str, Tool, Function]]
     frontend_state: Optional[ProgramState]
     grammar_sampling: bool
-
-
-def convert_tool_calls_to_function_call(
-    functions: Optional[List[Function]], chat_message: Dict
-) -> Dict:
-    if (
-        functions
-        and len(functions) > 0
-        and "tool_calls" in chat_message
-        and chat_message["tool_calls"] is not None
-        and len(chat_message["tool_calls"]) > 0
-    ):
-        chat_message["function_call"] = {
-            "name": chat_message["tool_calls"][0]["function"]["name"],
-            "arguments": chat_message["tool_calls"][0]["function"]["arguments"],
-        }
-        chat_message["tool_calls"] = None
-
-    return chat_message
 
 
 def v1_chat_generate_request(
@@ -382,19 +364,12 @@ async def completion_stream_generator(params: ChatCompletionParams):
         params.tools_or_functions,
     ):
         # Convert tool_calls to function_call if request.functions is provided
-        if (
-            params.request.functions
-            and len(params.request.functions) > 0
-            and "tool_calls" in response["delta"]
-            and response["delta"]["tool_calls"]
-            and len(response["delta"]["tool_calls"]) > 0
-        ):
-            tool_name = response["delta"]["tool_calls"][0]["function"]["name"]
-            tool_args = response["delta"]["tool_calls"][0]["function"]["arguments"]
-            response["delta"]["function_call"] = response["delta"]["tool_calls"][0][
-                "function"
-            ]
-            response["delta"]["tool_calls"] = None
+        response = convert_tool_calls_to_function_call(
+            functions=params.request.functions, chat_message=response
+        )
+        if response["delta"]["function_call"]:
+            tool_name = response["delta"]["function_call"]["name"]
+            tool_args = response["delta"]["function_call"]["arguments"]
             if tool_name and len(tool_name) > 0 and tool_args == "":
                 tool_call_count += 1
